Push tests (#87)
* added Pull and Push tests

* improved push_test

+++ src/NBomber/Domain/StatisticsTypes.fs
@@ -10,14 +10,14 @@ type LatencyCount = {
 }
 
 type Percentiles = {
-    RPS: int64
+    RPS: int
     Min: Latency
     Mean: Latency
     Max: Latency
     Percent50: Latency
     Percent75: Latency
     Percent95: Latency
-    StdDev: int64
+    StdDev: int
 }
 
 type StepStats = {
